requirements.txt
FastAPI
opentelemetry-api
opentelemetry-sdk
uvicorn
opentelemetry-distro
opentelemetry-exporter-otlp